requirements.txt

cohere==5.11.3
Flask==3.0.3
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
langchain==0.3.7
langchain-cohere==0.3.1
langchain-community==0.3.5
langchain-core==0.3.15
langchain-experimental==0.3.3
langchain-google-genai==2.0.4
langchain-postgres==0.0.12
langchain-text-splitters==0.3.2
pgvector==0.2.5
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
psycopg2-binary==2.9.10
PyMuPDF==1.24.13
python-dotenv==1.0.1